Optimizations for rebalance methods
Cache frequently accessed member variables in local variables.
It's likely that this allows the compiler to do a better job of
enregistering them, since it doesn't have to consider the potential for
aliasing when dealing with local variables whose address is never taken.

Regardless of the mechanism, the result is a modest but consistently
measurable improvement (g++ 13.3).